requirements.txt
requests>=2.31.0
dnspython>=2.4.0
phonenumbers>=8.13.0
cryptography>=41.0.0
flask>=3.0.0
click>=8.1.0
pyyaml>=6.0.1
pytest>=7.4.0
pytest-mock>=3.12.0
responses>=0.24.0
urllib3>=2.1.0
